MattTheLegoman's Update
Added Acacia Log, log_acacia & log_acacia_top, with fixes to contrast.
Version before this was not added with my knowledge or permission. You
need permission to add other people's stuff to a thing that is Creative
Commons, once it is is there I'm sure you know the rules for that (free
to do whatever). I'm still around to tell you off.

Fixed transparency issue in Lavandula plant.